refactor : RandomNumber 유틸 클래스 분리
