+++ src/main/java/study/race/CarRace.java
@@ -2,11 +2,10 @@ package study.race;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Random;
+
+import study.util.RandomNumber;
 
 public class CarRace {
-
-    private Random random;
 
     private List<Car> cars;
     private int tryNum;
@@ -28,7 +27,7 @@ public class CarRace {
     public void run() {
         for (int i = 0; i < tryNum; i++) {
             for (Car car : cars) {
-                int randomNum = getRandomNumber();
+                int randomNum = RandomNumber.getRandomNumber();
                 car.move(randomNum);
             }
         }
@@ -36,13 +35,6 @@ public class CarRace {
 
     public List<Car> getCars() {
         return cars;
-    }
-
-    private int getRandomNumber() {
-        if (random == null) {
-            random = new Random();
-        }
-        return random.nextInt(10);
     }
 
 }
