+++ src/main/java/study/util/RandomNumber.java
@@ -1,0 +1,13 @@
+package study.util;
+
+import java.util.Random;
+
+public class RandomNumber {
+    
+    private static final Random random = new Random();
+
+    public static int getRandomNumber() {
+        return random.nextInt(10);
+    }
+
+}
