+++ src/main/java/study/util/StringAddCalculator.java
@@ -1,4 +1,4 @@
-package study.calculator;
+package study.util;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
