+++ src/test/java/study/util/TestStringAddCalculator.java
@@ -1,11 +1,11 @@
-package study.calculator;
+package study.util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
 
-import study.calculator.StringAddCalculator;
+import study.util.StringAddCalculator;
 
 
 public class TestStringAddCalculator {
